# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ocean API",
    "version": "0.1.0",
    "description": "REST API for a Notion-like collaborative workspace backend."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development"
    },
    {
      "url": "https://ocean-api-269299350620.europe-west1.run.app",
      "description": "Cloud Run"
    }
  ],
  "security": [
    {
      "firebaseBearer": []
    }
  ],
...
  "paths": {
...
    "/api/workspaces/{workspaceId}/pages": {
      "get": {
        "tags": ["Pages"],
        "summary": "List all workspace pages",
        "parameters": [{ "$ref": "#/components/parameters/WorkspaceId" }],
        "responses": {
          "200": {
            "description": "Workspace pages",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Page" }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" }
        }
      },
      "post": {
        "tags": ["Pages"],
        "summary": "Create page inside workspace",
        "parameters": [{ "$ref": "#/components/parameters/WorkspaceId" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateWorkspacePageRequest" }
            }
          }
        },
        "responses": {
          "201": { "$ref": "#/components/responses/PageResponse" },
          "400": { "$ref": "#/components/responses/Error" },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
    "/api/pages": {
      "post": {
        "tags": ["Pages"],
        "summary": "Create page",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreatePageRequest" }
            }
          }
        },
        "responses": {
          "201": { "$ref": "#/components/responses/PageResponse" },
          "400": { "$ref": "#/components/responses/Error" },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/pages/{pageId}": {
      "get": {
        "tags": ["Pages"],
        "summary": "Get page",
        "parameters": [{ "$ref": "#/components/parameters/PageId" }],
        "responses": {
          "200": { "$ref": "#/components/responses/PageResponse" },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      },
      "patch": {
        "tags": ["Pages"],
        "summary": "Update page",
        "parameters": [{ "$ref": "#/components/parameters/PageId" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdatePageRequest" }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/PageResponse" },
          "400": { "$ref": "#/components/responses/Error" },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      },
      "put": {
        "tags": ["Pages"],
        "summary": "Update page",
        "parameters": [{ "$ref": "#/components/parameters/PageId" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdatePageRequest" }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/PageResponse" },
          "400": { "$ref": "#/components/responses/Error" },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      },
      "delete": {
        "tags": ["Pages"],
        "summary": "Soft delete page",
        "parameters": [{ "$ref": "#/components/parameters/PageId" }],
        "responses": {
          "204": { "description": "Deleted" },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/pages/{pageId}/content": {
      "get": {
        "tags": ["Pages"],
        "summary": "Get page with ordered blocks and plain text",
        "parameters": [{ "$ref": "#/components/parameters/PageId" }],
        "responses": {
          "200": { "$ref": "#/components/responses/PageContentResponse" },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/pages/{pageId}/title": {
      "patch": {
        "tags": ["Pages"],
        "summary": "Update page title",
        "description": "Dedicated title update route. Accepts either title or name.",
        "parameters": [{ "$ref": "#/components/parameters/PageId" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdatePageTitleRequest" }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/PageResponse" },
          "400": { "$ref": "#/components/responses/Error" },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/pages/{pageId}/restore": {
      "post": {
        "tags": ["Pages"],
        "summary": "Restore soft-deleted page",
        "parameters": [{ "$ref": "#/components/parameters/PageId" }],
        "responses": {
          "204": { "description": "Restored" },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/pages/{pageId}/move": {
      "post": {
        "tags": ["Pages"],
        "summary": "Move page",
        "parameters": [{ "$ref": "#/components/parameters/PageId" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MovePageRequest" }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/PageResponse" },
          "400": { "$ref": "#/components/responses/Error" },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/pages/{pageId}/children": {
      "get": {
        "tags": ["Pages"],
        "summary": "List child pages",
        "parameters": [{ "$ref": "#/components/parameters/PageId" }],
        "responses": {
          "200": {
            "description": "Child pages",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Page" }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "firebaseBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Firebase ID token"
      }
    },
    "parameters": {
      "WorkspaceId": {
        "name": "workspaceId",
        "in": "path",
        "required": true,
        "schema": { "type": "string" }
      },
      "PageId": {
        "name": "pageId",
        "in": "path",
        "required": true,
        "schema": { "type": "string" }
      },
...
    },
...
    "responses": {
      "Error": {
        "description": "Error response",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
...
      "PageResponse": {
        "description": "Page response",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "data": { "$ref": "#/components/schemas/Page" }
              }
            }
          }
        }
      },
      "PageContentResponse": {
        "description": "Page content response",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "properties": {
                    "page": { "$ref": "#/components/schemas/Page" },
                    "blocks": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Block" }
                    },
                    "plainText": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "message": { "type": "string" },
              "details": {}
            },
            "required": ["message"]
          }
        },
        "required": ["error"]
      },
...
      "Page": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "workspaceId": { "type": "string" },
          "parentPageId": { "type": "string", "nullable": true },
          "title": { "type": "string" },
          "icon": { "type": "string" },
          "coverImage": { "type": "string", "nullable": true },
          "createdBy": { "type": "string" },
          "updatedBy": { "type": "string" },
          "visibility": { "type": "string", "enum": ["private", "workspace", "public"] },
          "order": { "type": "number" },
          "isDeleted": { "type": "boolean" },
          "deletedAt": { "$ref": "#/components/schemas/TimestampNullable" },
          "createdAt": { "$ref": "#/components/schemas/Timestamp" },
          "updatedAt": { "$ref": "#/components/schemas/Timestamp" }
        }
      },
      "CreatePageRequest": {
        "type": "object",
        "properties": {
          "workspaceId": { "type": "string" },
          "parentPageId": { "type": "string", "nullable": true },
          "parentId": { "type": "string", "nullable": true, "description": "Alias for parentPageId" },
          "title": { "type": "string", "default": "Untitled", "maxLength": 240 },
          "name": { "type": "string", "maxLength": 240, "description": "Alias for title" },
          "icon": { "type": "string", "default": "page", "maxLength": 16 },
          "coverImage": { "type": "string", "format": "uri", "nullable": true },
          "visibility": { "type": "string", "enum": ["private", "workspace", "public"], "default": "workspace" },
          "order": { "type": "number" }
        },
        "required": ["workspaceId"]
      },
      "CreateWorkspacePageRequest": {
        "type": "object",
        "description": "Create page request for /api/workspaces/{workspaceId}/pages. workspaceId comes from the path.",
        "properties": {
          "parentPageId": { "type": "string", "nullable": true },
          "parentId": { "type": "string", "nullable": true, "description": "Alias for parentPageId" },
          "title": { "type": "string", "default": "Untitled", "maxLength": 240 },
          "name": { "type": "string", "maxLength": 240, "description": "Alias for title" },
          "icon": { "type": "string", "default": "page", "maxLength": 16 },
          "coverImage": { "type": "string", "format": "uri", "nullable": true },
          "visibility": { "type": "string", "enum": ["private", "workspace", "public"], "default": "workspace" },
          "order": { "type": "number" }
        }
      },
      "UpdatePageRequest": {
        "type": "object",
        "properties": {
          "title": { "type": "string", "maxLength": 240 },
          "name": { "type": "string", "maxLength": 240, "description": "Alias for title" },
          "icon": { "type": "string", "maxLength": 16 },
          "coverImage": { "type": "string", "format": "uri", "nullable": true },
          "visibility": { "type": "string", "enum": ["private", "workspace", "public"] },
          "order": { "type": "number" }
        }
      },
      "UpdatePageTitleRequest": {
        "type": "object",
        "description": "Send either title or name.",
        "properties": {
          "title": { "type": "string", "maxLength": 240 },
          "name": { "type": "string", "maxLength": 240, "description": "Alias for title" }
        }
      },
...
      "MovePageRequest": {
        "type": "object",
        "properties": {
          "parentPageId": { "type": "string", "nullable": true },
          "order": { "type": "number" }
        },
        "required": ["parentPageId"]
      },
      "Block": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "pageId": { "type": "string" },
          "workspaceId": { "type": "string" },
          "type": { "$ref": "#/components/schemas/BlockType" },
          "content": { "type": "object", "additionalProperties": true },
          "parentBlockId": { "type": "string", "nullable": true },
          "order": { "type": "number" },
          "createdBy": { "type": "string" },
          "updatedBy": { "type": "string" },
          "isDeleted": { "type": "boolean" },
          "createdAt": { "$ref": "#/components/schemas/Timestamp" },
          "updatedAt": { "$ref": "#/components/schemas/Timestamp" }
        }
      },
...
      "BlockType": {
        "type": "string",
        "enum": [
          "paragraph",
          "heading_1",
          "heading_2",
          "heading_3",
          "bulleted_list",
          "numbered_list",
          "todo",
          "toggle",
          "quote",
          "callout",
          "divider",
          "code",
          "image",
          "video",
          "audio",
          "file",
          "bookmark",
          "equation",
          "table",
          "columns",
          "database_view",
          "embed",
          "breadcrumb",
          "table_of_contents",
          "synced_block"
        ]
      },
...
      "Timestamp": {
        "oneOf": [
          { "type": "string", "format": "date-time" },
          { "type": "object", "additionalProperties": true }
        ]
      },
      "TimestampNullable": {
        "nullable": true,
        "oneOf": [
          { "type": "string", "format": "date-time" },
          { "type": "object", "additionalProperties": true }
        ]
      }
    }
  }
}